+++ frontend/javascripts/libs/async/latest_task_executor.ts
@@ -1,4 +1,4 @@
-import Deferred from "libs/deferred";
+import Deferred from "libs/async/deferred";
 type Task<T> = () => Promise<T>;
 export const SKIPPED_TASK_REASON = "Skipped task";
 /*
