+++ frontend/javascripts/oxalis/model/sagas/isosurface_saga.ts
@@ -5,7 +5,7 @@ import { chunkDynamically, sleep } from "libs/utils";
 import ErrorHandling from "libs/error_handling";
 import type { APIDataset, APIMeshFile, APISegmentationLayer } from "types/api_flow_types";
 import { mergeBufferGeometries, mergeVertices } from "libs/BufferGeometryUtils";
-import Deferred from "libs/deferred";
+import Deferred from "libs/async/deferred";
 
 import Store from "oxalis/store";
 import {
@@ -70,7 +70,7 @@ import { saveNowAction } from "oxalis/model/actions/save_actions";
 import Toast from "libs/toast";
 import { getDracoLoader } from "libs/draco";
 import messages from "messages";
-import processTaskWithPool from "libs/task_pool";
+import processTaskWithPool from "libs/async/task_pool";
 import { getBaseSegmentationName } from "oxalis/view/right-border-tabs/segments_tab/segments_view_helper";
 import { RemoveSegmentAction, UpdateSegmentAction } from "../actions/volumetracing_actions";
 import { ResolutionInfo } from "../helpers/resolution_info";
