+++ frontend/javascripts/test/libs/async_fifo_resolver.spec.ts
@@ -1,7 +1,7 @@
 import { sleep } from "libs/utils";
 import test from "ava";
 import _ from "lodash";
-import { AsyncFifoResolver } from "libs/async_fifo_resolver";
+import { AsyncFifoResolver } from "libs/async/async_fifo_resolver";
 
 const createSubmitterFnWithProtocol = () => {
   const resolver = new AsyncFifoResolver();
