+++ frontend/javascripts/test/libs/debounced_abortable_saga.spec.ts
@@ -2,7 +2,7 @@ import { type Saga } from "oxalis/model/sagas/effect-generators";
 import { sleep } from "libs/utils";
 import test from "ava";
 import _ from "lodash";
-import { createDebouncedAbortableCallable } from "libs/debounced_abortable_saga";
+import { createDebouncedAbortableCallable } from "libs/async/debounced_abortable_saga";
 
 const createAbortableFnWithProtocol = () => {
   const protocol: string[] = [];
