+++ frontend/javascripts/test/libs/deferred.spec.ts
@@ -1,5 +1,5 @@
 // @ts-nocheck
-import Deferred from "libs/deferred";
+import Deferred from "libs/async/deferred";
 import runAsync from "test/helpers/run-async";
 import test from "ava";
