+++ frontend/javascripts/test/libs/task_pool.spec.ts
@@ -1,6 +1,6 @@
 import { call, type Saga } from "oxalis/model/sagas/effect-generators";
 import { runSaga } from "redux-saga";
-import processTaskWithPool from "libs/task_pool";
+import processTaskWithPool from "libs/async/task_pool";
 import * as Utils from "libs/utils";
 import test from "ava";
